Add descriptions module
