+++ homeassistant/components/tplink/diagnostics.py
@@ -5,12 +5,10 @@ from __future__ import annotations
 from typing import Any
 
 from homeassistant.components.diagnostics import async_redact_data
-from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.device_registry import format_mac
 
-from .const import DOMAIN
-from .models import TPLinkData
+from . import TPLinkConfigEntry
 
 TO_REDACT = {
     # Entry fields
@@ -42,10 +40,10 @@ TO_REDACT = {
 
 
 async def async_get_config_entry_diagnostics(
-    hass: HomeAssistant, entry: ConfigEntry
+    hass: HomeAssistant, entry: TPLinkConfigEntry
 ) -> dict[str, Any]:
     """Return diagnostics for a config entry."""
-    data: TPLinkData = hass.data[DOMAIN][entry.entry_id]
+    data = entry.runtime_data
     coordinator = data.parent_coordinator
     oui = format_mac(coordinator.device.mac)[:8].upper()
     return async_redact_data(
