+++ homeassistant/components/tplink/sensor.py
@@ -2,122 +2,18 @@
 
 from __future__ import annotations
 
-from dataclasses import dataclass
-from datetime import datetime
 from typing import cast
 
 from kasa import Device, Feature
 
-from homeassistant.components.sensor import (
-    SensorDeviceClass,
-    SensorEntity,
-    SensorEntityDescription,
-    SensorStateClass,
-)
-from homeassistant.const import (
-    ATTR_VOLTAGE,
-    UnitOfElectricCurrent,
-    UnitOfElectricPotential,
-    UnitOfEnergy,
-    UnitOfPower,
-)
+from homeassistant.components.sensor import SensorEntity
 from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
 from . import TPLinkConfigEntry
-from .const import (
-    ATTR_CURRENT_A,
-    ATTR_CURRENT_POWER_W,
-    ATTR_TODAY_ENERGY_KWH,
-    ATTR_TOTAL_ENERGY_KWH,
-)
 from .coordinator import TPLinkDataUpdateCoordinator
-from .entity import (
-    CoordinatedTPLinkEntity,
-    TPLinkFeatureEntityDescription,
-    _category_for_feature,
-    _entities_for_device_and_its_children,
-)
-
-
-@dataclass(frozen=True)
-class TPLinkSensorEntityDescription(
-    SensorEntityDescription, TPLinkFeatureEntityDescription
-):
-    """Describes TPLink sensor entity."""
-
-    precision: int | None = None
-
-
-ENERGY_SENSORS: tuple[TPLinkSensorEntityDescription, ...] = (
-    TPLinkSensorEntityDescription(
-        key=ATTR_CURRENT_POWER_W,
-        translation_key="current_consumption",
-        native_unit_of_measurement=UnitOfPower.WATT,
-        device_class=SensorDeviceClass.POWER,
-        state_class=SensorStateClass.MEASUREMENT,
-        feature_id="current_consumption",
-        precision=1,
-    ),
-    TPLinkSensorEntityDescription(
-        key=ATTR_TOTAL_ENERGY_KWH,
-        translation_key="total_consumption",
-        native_unit_of_measurement=UnitOfEnergy.KILO_WATT_HOUR,
-        device_class=SensorDeviceClass.ENERGY,
-        state_class=SensorStateClass.TOTAL_INCREASING,
-        feature_id="consumption_total",
-        precision=3,
-    ),
-    TPLinkSensorEntityDescription(
-        key=ATTR_TODAY_ENERGY_KWH,
-        translation_key="today_consumption",
-        native_unit_of_measurement=UnitOfEnergy.KILO_WATT_HOUR,
-        device_class=SensorDeviceClass.ENERGY,
-        state_class=SensorStateClass.TOTAL_INCREASING,
-        feature_id="consumption_today",
-        precision=3,
-    ),
-    TPLinkSensorEntityDescription(
-        key=ATTR_VOLTAGE,
-        native_unit_of_measurement=UnitOfElectricPotential.VOLT,
-        device_class=SensorDeviceClass.VOLTAGE,
-        state_class=SensorStateClass.MEASUREMENT,
-        feature_id="voltage",
-        precision=1,
-    ),
-    TPLinkSensorEntityDescription(
-        key=ATTR_CURRENT_A,
-        native_unit_of_measurement=UnitOfElectricCurrent.AMPERE,
-        device_class=SensorDeviceClass.CURRENT,
-        state_class=SensorStateClass.MEASUREMENT,
-        feature_id="current",
-        precision=2,
-    ),
-)
-
-
-def _new_sensor_description(feature: Feature) -> TPLinkSensorEntityDescription:
-    """Create description for entities not yet added to the static description list."""
-
-    if isinstance(feature.value, datetime):
-        device_class = SensorDeviceClass.TIMESTAMP
-        state_class = None
-        native_unit_of_measurement = None
-    else:
-        device_class = None
-        state_class = SensorStateClass.MEASUREMENT
-        native_unit_of_measurement = feature.unit
-    return TPLinkSensorEntityDescription(
-        key=feature.id,
-        name=feature.name,
-        device_class=device_class,
-        state_class=state_class,
-        native_unit_of_measurement=native_unit_of_measurement,
-        feature_id=feature.id,
-        precision=feature.precision_hint,
-        entity_registry_enabled_default=feature.category is not Feature.Category.Debug,
-        entity_category=_category_for_feature(feature),
-    )
+from .descriptions import TPLinkSensorEntityDescription
+from .entity import CoordinatedTPLinkEntity, _entities_for_device_and_its_children
 
 
 async def async_setup_entry(
@@ -137,9 +33,7 @@ async def async_setup_entry(
         coordinator=parent_coordinator,
         feature_type=Feature.Type.Sensor,
         entity_class=TPLinkSensor,
-        descriptions=ENERGY_SENSORS,
         child_coordinators=children_coordinators,
-        new_description_generator=_new_sensor_description,
     )
     async_add_entities(entities)
