+++ homeassistant/components/tplink/switch.py
@@ -2,45 +2,25 @@
 
 from __future__ import annotations
 
-from dataclasses import dataclass
 import logging
 from typing import Any
 
 from kasa import Device, Feature
 
-from homeassistant.components.switch import SwitchEntity, SwitchEntityDescription
+from homeassistant.components.switch import SwitchEntity
 from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
 from . import TPLinkConfigEntry
 from .coordinator import TPLinkDataUpdateCoordinator
+from .descriptions import TPLinkSwitchEntityDescription
 from .entity import (
     CoordinatedTPLinkEntity,
-    TPLinkFeatureEntityDescription,
-    _category_for_feature,
     _entities_for_device_and_its_children,
     async_refresh_after,
 )
 
 _LOGGER = logging.getLogger(__name__)
-
-
-@dataclass(frozen=True)
-class TPLinkSwitchEntityDescription(
-    SwitchEntityDescription, TPLinkFeatureEntityDescription
-):
-    """Describes TPLink switch entity."""
-
-
-def _new_switch_description(feature: Feature) -> TPLinkSwitchEntityDescription:
-    """Create description for entities not yet added to the static description list."""
-    return TPLinkSwitchEntityDescription(
-        key=feature.id,
-        name=feature.name,
-        feature_id=feature.id,
-        entity_registry_enabled_default=feature.category is not Feature.Category.Debug,
-        entity_category=_category_for_feature(feature),
-    )
 
 
 async def async_setup_entry(
@@ -58,7 +38,6 @@ async def async_setup_entry(
         coordinator=parent_coordinator,
         feature_type=Feature.Switch,
         entity_class=TPLinkSwitch,
-        new_description_generator=_new_switch_description,
     )
 
     async_add_entities(entities)
